[refactor] Delegating building descriptors to the Policy impl
* Extracted to `crate::policy::Policy`
