+++ src/filter/http_context.rs
@@ -1,11 +1,7 @@
-use crate::configuration::{
-    Condition, DataItem, DataType, FailureMode, FilterConfig, PatternExpression, Policy, Rule,
-};
-use crate::envoy::{
-    RateLimitDescriptor, RateLimitDescriptor_Entry, RateLimitRequest, RateLimitResponse,
-    RateLimitResponse_Code,
-};
+use crate::configuration::{FailureMode, FilterConfig};
+use crate::envoy::{RateLimitRequest, RateLimitResponse, RateLimitResponse_Code};
 use crate::filter::http_context::TracingHeader::{Baggage, Traceparent, Tracestate};
+use crate::policy::Policy;
 use log::{debug, warn};
 use protobuf::Message;
 use proxy_wasm::traits::{Context, HttpContext};
@@ -59,7 +55,7 @@ impl Filter {
     }
 
     fn process_rate_limit_policy(&self, rlp: &Policy) -> Action {
-        let descriptors = self.build_descriptors(rlp);
+        let descriptors = rlp.build_descriptors(self);
         if descriptors.is_empty() {
             debug!(
                 "#{} process_rate_limit_policy: empty descriptors",
@@ -104,114 +100,6 @@ impl Filter {
                 Action::Continue
             }
         }
-    }
-
-    fn build_descriptors(&self, rlp: &Policy) -> protobuf::RepeatedField<RateLimitDescriptor> {
-        rlp.rules
-            .iter()
-            .filter(|rule: &&Rule| self.filter_rule_by_conditions(&rule.conditions))
-            // Mapping 1 Rule -> 1 Descriptor
-            // Filter out empty descriptors
-            .filter_map(|rule| self.build_single_descriptor(&rule.data))
-            .collect()
-    }
-
-    fn filter_rule_by_conditions(&self, conditions: &[Condition]) -> bool {
-        if conditions.is_empty() {
-            // no conditions is equivalent to matching all the requests.
-            return true;
-        }
-
-        conditions
-            .iter()
-            .any(|condition| self.condition_applies(condition))
-    }
-
-    fn condition_applies(&self, condition: &Condition) -> bool {
-        condition
-            .all_of
-            .iter()
-            .all(|pattern_expression| self.pattern_expression_applies(pattern_expression))
-    }
-
-    fn pattern_expression_applies(&self, p_e: &PatternExpression) -> bool {
-        let attribute_path = p_e.path();
-        let attribute_value = match self.get_property(attribute_path) {
-            None => {
-                debug!(
-                    "#{} pattern_expression_applies:  selector not found: {}, defaulting to ``",
-                    self.context_id, p_e.selector
-                );
-                b"".to_vec()
-            }
-            Some(attribute_bytes) => attribute_bytes,
-        };
-        match p_e.eval(attribute_value) {
-            Err(e) => {
-                debug!(
-                    "#{} pattern_expression_applies failed: {}",
-                    self.context_id, e
-                );
-                false
-            }
-            Ok(result) => result,
-        }
-    }
-
-    fn build_single_descriptor(&self, data_list: &[DataItem]) -> Option<RateLimitDescriptor> {
-        let mut entries = ::protobuf::RepeatedField::default();
-
-        // iterate over data items to allow any data item to skip the entire descriptor
-        for data in data_list.iter() {
-            match &data.item {
-                DataType::Static(static_item) => {
-                    let mut descriptor_entry = RateLimitDescriptor_Entry::new();
-                    descriptor_entry.set_key(static_item.key.to_owned());
-                    descriptor_entry.set_value(static_item.value.to_owned());
-                    entries.push(descriptor_entry);
-                }
-                DataType::Selector(selector_item) => {
-                    let descriptor_key = match &selector_item.key {
-                        None => selector_item.path().to_string(),
-                        Some(key) => key.to_owned(),
-                    };
-
-                    let attribute_path = selector_item.path();
-                    let value = match self.get_property(attribute_path.tokens()) {
-                        None => {
-                            debug!(
-                                "#{} build_single_descriptor: selector not found: {}",
-                                self.context_id, attribute_path
-                            );
-                            match &selector_item.default {
-                                None => return None, // skipping the entire descriptor
-                                Some(default_value) => default_value.clone(),
-                            }
-                        }
-                        // TODO(eastizle): not all fields are strings
-                        // https://www.envoyproxy.io/docs/envoy/latest/intro/arch_overview/advanced/attributes
-                        Some(attribute_bytes) => match String::from_utf8(attribute_bytes) {
-                            Err(e) => {
-                                debug!(
-                                    "#{} build_single_descriptor: failed to parse selector value: {}, error: {}",
-                                    self.context_id, attribute_path, e
-                                );
-                                return None;
-                            }
-                            Ok(attribute_value) => attribute_value,
-                        },
-                    };
-                    let mut descriptor_entry = RateLimitDescriptor_Entry::new();
-                    descriptor_entry.set_key(descriptor_key);
-                    descriptor_entry.set_value(value);
-                    entries.push(descriptor_entry);
-                }
-            }
-        }
-
-        let mut res = RateLimitDescriptor::new();
-        res.set_entries(entries);
-        Some(res)
     }
 
     fn handle_error_on_grpc_response(&self) {
