+++ src/lib.rs
@@ -2,6 +2,7 @@ mod configuration;
 mod envoy;
 mod filter;
 mod glob;
+mod policy;
 mod policy_index;
 
 #[cfg(test)]
